address validation, address contained in the wallet, layout sw600dp support, tx saving, tx loading from sqlite to wallet

+++ app/src/main/java/com/breadwallet/wallet/BRPeerManager.java
@@ -52,7 +52,7 @@ public class BRPeerManager {
         return instance;
     }
 
-    public native void connect(long earliestKeyTime, long blockCount, long peerCount);
+    public native void connect(long earliestKeyTime, int blockCount, int peerCount);
 
     public native void putPeer(byte[] peerAddress, byte[] peerPort, byte[] peerTimeStamp);
